% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotype.R
\name{GetGenotype}
\alias{GetGenotype}
\title{Title}
\usage{
GetGenotype(model_path, M, S)
}
\arguments{
\item{model_path}{s}

\item{M}{s}

\item{S}{s}

\item{model_lattice_path}{s}

\item{modelPrior}{s}
}
\description{
Title
}
