% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMandS.R
\name{GetMandS}
\alias{GetMandS}
\title{Get M and S from raw reads}
\usage{
GetMandS(
  snps_path,
  bigWig_path,
  coverage_cutoff = 4,
  alt_path,
  sample_id_rep,
  temp_folder,
  accuracyModel
)
}
\arguments{
\item{snps_path}{Path to biallelic SNPs to be genotyped.}

\item{bigWig_path}{Path to the bigWig file containing the total read counts for a single sample.}

\item{coverage_cutoff}{The minimum amount of read count mapped to a loci for that SNP to be included in the genotype calling.}

\item{alt_path}{Path to alternative base read counts. This is outputed from the Recount3 pipeline as a .zst file.}

\item{sample_id_rep}{Single sample ID to be genotyped.}

\item{temp_folder}{Path to temporary folder.}

\item{accuracyModel}{loaded accuracy model}
}
\description{
This function will generate the reference and alternative base counts for each loci. The alternative base input is the custom file format (.zst file) from
the Recount3. This file includes read counts for the 4 bases. Therefore, in this function, we will find the base that has read counts
mapped to it. The default coverage cutoff is set to 4. After the ref and alt counts are obtained, the M and S values will be calculated and returened.
}
