% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAltCount.R
\name{get_alt_count}
\alias{get_alt_count}
\title{Title}
\usage{
get_alt_count(alt_path, sample_id_rep, filtered_snps_gr, temp_folder)
}
\arguments{
\item{alt_path}{The path that leads to the Recount3 alt.zst file}

\item{sample_id_rep}{The individual sample ID to be genotyped}

\item{filtered_snps_gr}{The biallelic SNP loci}

\item{temp_folder}{Temporary folder to save unwanted files}
}
\description{
Title
}
