% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotype.R
\name{GetGenotype}
\alias{GetGenotype}
\title{Title}
\usage{
GetGenotype(
  model_MS,
  modelLattice,
  prior = c(0.93752717, 0.03951271, 0.02239503),
  M,
  S
)
}
\arguments{
\item{model_MS}{Genotyping model}

\item{modelLattice}{A lookup table generated form the genotyping model for faster performance}

\item{M}{M values calculated from GetMandS function}

\item{S}{S value calculated from GetMandS fucntion}
}
\description{
Title
}
