% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotype.R
\name{GetGenotype}
\alias{GetGenotype}
\title{Predict the sample genotype using M and S values}
\usage{
GetGenotype(model, prior = c(0.93752717, 0.03951271, 0.02239503), M, S)
}
\arguments{
\item{model}{Genotyping model}

\item{M}{M values calculated from GetMandS function}

\item{S}{S value calculated from GetMandS function}
}
\value{
array of the predicted genotype. The SNP order is the same as output from GetMandS function.
}
\description{
This function will use the M and S values generated from the GetMandS function to predict the genotype.
}
