% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotype.R
\name{GetGenotype}
\alias{GetGenotype}
\title{A function for calculating genotypes on the basis of M an S values}
\usage{
GetGenotype(model = NULL, prior = c(0.93752717, 0.03951271, 0.02239503), M, S)
}
\arguments{
\item{model}{Our genotyping model. Our model uses linear regression and GAM to model the mean and variance relationship between M and S values. Our model is a list object containing linear and GAM model for the 3 genotypes.}

\item{prior}{prior was calculated based on the 3 genotype distribution in our training set and default is c(0.93752717, 0.03951271, 0.02239503). This value is constant in our model.}

\item{M}{M values calculated from GetMandS function}

\item{S}{S value calculated from GetMandS function}
}
\value{
array of the predicted genotype. The SNP order is the same as output from GetMandS function.
}
\description{
This function will use the M and S values generated from the GetMandS function to predict the genotype.
Using our genotyping model, this function predicts the genotype: 1 (homozygous reference), 2 (heterozygous), 3 (homozygous alternative)
}
\examples{

S=test$S
M=test$M

test$predicted_genotype<-GetGenotype(model, M, S)


}
