% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenotype.R
\name{GetGenotype}
\alias{GetGenotype}
\title{Title}
\usage{
GetGenotype(model, prior = c(0.93752717, 0.03951271, 0.02239503), M, S)
}
\arguments{
\item{model}{Genotyping model}

\item{M}{M values calculated from GetMandS function}

\item{S}{S value calculated from GetMandS funCtion}
}
\description{
Title
}
