% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMandS.R
\name{GetMandS}
\alias{GetMandS}
\title{Get M and S from raw reads as well as the predicted accuracy for that SNP}
\usage{
GetMandS(
  snps,
  bigWig_path,
  coverage_cutoff = 4,
  alt_path,
  sample_id_rep,
  temp_folder
)
}
\arguments{
\item{snps}{The biallelic SNPs to be genotyped. This file should be a grange object containing seqnames,ranges,strand,ref_seq,alt_seq,allele_freq}

\item{bigWig_path}{Path to the bigWig file containing the total read counts for a single sample.}

\item{coverage_cutoff}{The minimum amount of read count mapped to a loci for that SNP to be included in the genotype calling. The default is 4 which was used in our model training and testing.}

\item{alt_path}{Path to alternative base read counts. This is outputed from the Recount3 pipeline as a .zst file.}

\item{sample_id_rep}{Single sample ID to be genotyped.}

\item{temp_folder}{Path to temporary folder.}
}
\description{
This function will generate the reference and alternative base counts for each loci. The alternative base input is the custom file format (.zst file) from
the Recount3. This file includes read counts for the 4 bases. Therefore, in this function, we will find the base that has read counts
mapped to it. The default coverage cutoff is set to 4. After the ref and alt counts are obtained, the M and S values will be calculated and returened.
}
\examples{

base_path<-system.file("extdata", package="RecountGenotyper")
alt_path=paste0(base_path,"/test.csv.zst")
bigWig_path=paste0(base_path,"/test.bw")

sample_id_rep="test"
temp_folder="~"

test_geno<-GetMandS(snps_path, bigWig_path, alt_path, sample_id_rep, temp_folder)

}
